inc: allegro5: Fix Win32 #inclibs
For Win32, the Allegro5 libs follow a specific naming pattern,
e.g. "liballegro-5.0.10-md.a":
     https://www.allegro.cc/manual/5/install/windows.html

On Ubuntu though it's just "liballegro.so".
We want to make #inclibs match that.

Even though the allegro5 binding is for 5.0.11, these #inclibs use 5.0.10,
because that seems to be the latest version of the Win32 binary builds
available from the Allegro homepage. They should be compatible anyways.