inc: New TRE (regex) binding
Besides updating binding for a more up-to-date TRE version, this also fixes
an issue where regex.bi used the libtre API/ABI but linked to glibc's
implementation of the reg*() functions, which are not ABI-compatible,
which caused the TRE examples to crash on Linux.

The new binding avoids this issue by using libtre's tre_reg*() functions -
thanks to the tre_* prefix, there is no collision with glibc's plain
functions anymore. In fact it seems like libtre doesn't expose plain
functions itself anymore, but only ones with the tre_* prefix, so I think
the old binding couldn't even have worked anymore on Win32 with the latest
libtre version.