inc: Add two OpenGL bindings: one Mesa, one Windows (MinGW-w64)
On GNU/Linux, the headers from Mesa are the standard. For MinGW however,
MinGW.org and MinGW-w64 provide specialized headers. They have mostly the
exact same declarations, except that the main GL/gl.h is limited to OpenGL
1.1, and anything more recent is treated as extension in GL/glext.h. With
Mesa that's not the case though.

Thus I think it's best to provide both header sets for FB, and default to
using the MinGW-w64 based ones on Windows, and the Mesa-based ones
otherwise.

As for Windows, I think the headers are mostly the same as before fc91503,
but for Linux the headers will now provide additional declarations
(those extensions that aren't directly available on Windows).