Rename FB_COMPOPT_DEBUG => FB_COMPOPT_DEBUGINFO
This should clarify that most code in the compiler that checks for this
option just wants to know whether debug info should be added or not.

This makes it easier to add other, related options in the future, e.g. for
separating the generation of debug info from the enabling of assertions.