rtlib/gfxlib2: Free main-thread-specific TLS data on program exit
Previously, the TLS contexts of the main thread were freed by fb_TlsExit(),
but since that's only used in the MT build, the TLS contexts were leaked
in the non-MT build.

Now we explicitly call fb_TlsFreeCtxTb() before fb_TlsExit(), in both
builds, to free the TLS contexts of the main thread.