Don't emit compilation time/date into output files
I think it can be useful to track the .asm/.c files generated by fbc in
Git, for example to have a precompiled version of fbc itself available,
or for testing fbc by checking for changes in the code generated by fbc.

For this it's a bit annoying to have time/date strings in the generated
files, because they change with every compilation. That unnecessarily
causes differences when comparing them to files generated previously.

Besides that, I don't think it's very useful to have time/date in there
anyways. It'd probably be more interesting to emit information about the
used compiler version, and the host/target systems...