makefile: Add bootstrap-dist and bootstrap commands
The idea is to allow precompiling fbc sources and packaging them up with
the FB souce tree into an FB source package that can be built without
needing an existing fbc.

The FreeBASIC-x.xx.x-source-bootstrap tarball could be uploaded to the fbc
download sites, next to the release packages.

Examples where it could be useful:
- packaging fbc in Linux distros which don't like importing binaries
- getting a working fbc on Linux distro A, where the fbc project's Linux
  binaries don't work because they were made on the incompatible Linux
  distro B
- getting a working fbc if you lose yours