inc: winapi: Re-add REFIID & co
These types are documented on MSDN so they shouldn't be dropped.

In the MS and in the MinGW-w64 headers they are #defines (not typedefs)
and have 2 CONSTs, so I have translated them that way for FB too.
But this doesn't match the typedefs shown in the MSDN documentation,
so I'm not 100% sure what the best solution is.