Add fbc build-time option to use lib64/ subdir for 64bit libs
(closes #742)

Some Linux distros want 64bit libs in lib64/ instead of lib/ (the latter
is probably reserved for 32bit then). To do this, I think it's best to add
this build-time configuration option.

It's not clear how other cross-compiling targets (besides 32bit x86 and
64bit x86_64) should be handled in this scenario. For now, lib/ will still
hold libs for x86 and arm, and lib64/ will still hold libs for x86_64 and
aarch64, etc. (but in practice this most likely doesn't matter, because the
distro will only package x86 and x86_64 builds anyways)