Stop using gcc target ids in lib/freebasic/<target> path
Previously,
	fbc -target i686-w64-mingw32
would use
	lib/freebasic/i686-w64-mingw32/
as the lib dir. This changes it to use
	lib/freebasic/win32/
i.e. the FB standard.

This means that cross-compiling (-target i686-w64-mingw32) and native
compilation (on win32) will use the same dir layout, making packaging and
distribution easier.

The only advantage of using the gcc target ids was that it allows doing
things like this:
	lib/freebasic/i486-pc-linux-gnu
	lib/freebasic/i686-pc-linux-gnu
which isn't possible with the <fbtarget> because that's just linux-x86
and doesn't encode the difference between i484 and i686. But I think it's
better to find a different solution for that problem, if needed, in
particular one that also works with the standalone dir layout (which only
uses the <fbtarget> ids).