inc: win/winuser.bi: Re-add ENUMWINDOWSPROC type
http://www.freebasic.net/forum/viewtopic.php?f=3&t=23488

fbc's old Windows API headers had the ENUMWINDOWSPROC type and used it in
some places instead of WNDENUMPROC. However, it's not documented on MSDN
and doesn't exist in MS or MinGW-w64 headers. Thus it was lost in the
automated translation of the new headers. We should probably re-add it
for backwards-compatibility, as it's easy and it seems that people used it.