Fix temp var destruction in static initializers
For static vars:
	static x as ClassUdt = ClassUdt( 123 )
we produce code like this:
	static inited as integer
	if inited = FALSE then
		inited = TRUE
		x.constructor( 123 )
	end if
i.e. the initialization code is wrapped in an If block so the variable is
only initialized once.

This did not take temp vars in the initializer expression into account;
they were destroyed by the astBuildBranch() in hWrapInStaticFlag(),
resulting in code like this:
	static inited as integer
	dim temp as ...
	condition = (inited = FALSE)
	temp.destructor()
	if condition then
		inited = TRUE
		temp.constructor(...)
		x.constructor( temp )
	end if

This is obviously broken since the temp var is destroyed too early, before
even being initialized.

To fix this, I adjusted the astBuildBranch() call to no longer call the
dtors, and instead, astDtorListFlush() is now done manually on the
initialization code path. Furthermore, cVarDecl() needs to use LINKs
and a single astAdd() instead of multiple astAdd()s for adding the DECL
node and the initialization code.

Besides initializer expressions, there can also be array bounds expressions
with temp vars in a static var decl (in case of static dynamic arrays),
this also works properly now. (hWrapInStaticFlag() is also used for the
Redim calls used to initialize the static dynamic arrays)