C backend: Fix pointer array access
http://www.freebasic.net/forum/viewtopic.php?f=17&t=23562

The C backend didn't consider that dynamic arrays will be emitted as
descriptors. It treated them as objects with the array element's type,
which could fool the code generation if the array element type was a
pointer. (it assumed the descriptor to be the pointer, instead of properly
accessing the descriptor's data pointer)

The array accesses worked with other (non-pointer) array element types,
but this only due to "luck".