internal: Rename FB_SYMBATTRIB_RETURNSBYREF to be more generic
so it can be re-used for Dim Byref, i.e. to mark any variables as
references. Ideally FB_SYMBATTRIB_PARAMBYREF would be removed in favour of
FB_SYMBATTRIB_REF too, but that's not as easy to solve out since
FB_SYMBATTRIB_PARAMBYREF carries additional information (besides the
BYREF), specifically that the symbol is a paramvar.