C backend: Fix handling of bitfields in nested anonymous structs
Because fbc's/gcc's bitfield layout differs currently, the C backend can't
emit bitfields in structs as-is, but has to emit such structs as a byte
array, to ensure that the sizeof() is what FB expects. See: 791b9578

This was only working if the Type had bitfields at the toplevel. Bitfields
in nested anonymous Types weren't handled. So now the FB_UDTOPT_HASBITFIELD
will be propagated up and it should work.