len/sizeof: Warn about ambigious identifiers
http://www.freebasic.net/forum/viewtopic.php?f=17&t=11567
http://www.freebasic.net/forum/viewtopic.php?f=3&t=23624

FB allows types and non-type symbols to have the same identifiers,
much like C. This is usually ok, because the language syntax doesn't
allow types and procedures/variables to be referenced by name in the
same places.

Except for len/sizeof/typeof, which (silently) resolve such ambiguities
by preferring the type. However, this can be quite surprising to the
programmer, for example if they are trying to get the len() of a local
string variable, or the sizeof() some local variable, but the compiler
silently returns the sizeof() some type defined in a header file far
away.

A common example involves the MSG type from the Windows API headers.
As it happens, "msg" is a fairly common variable name aswell when doing
Windows API programming.

	type MSG
		i as integer
	end type
	dim msg as string
	print len(msg) '' 4, not 0

To fix this, it seems best to add a warning about such ambigious cases.
Perhaps it would be more perfect to change the behaviour of the language,
for example:
  a) disallowing types/non-types with the same identifier
  b) allow len() for variables/expressions only, sizeof() on types only
  c) allow len() on string expressions only, sizeof() on anything (?)
  d) at least make len() prefer strings
but all of these ideas break compatibility or couldn't be agreed upon so
far, or don't fully solve the problem anyways.