Hide "ambigious sizeof" warning if id could only refer to type or procedure
It's not possible to do sizeof() on procedures anyways (it will result in
a syntax error), so there's no ambiguity here. We'll always use the type,
and that's probably what the user expected too, because functions can only
legally appear together with other tokens (in a function call, or an
address-of operation).