Reword "ambigious sizeof" warning if preferring var over fwdref
The idea is that sometimes we'll actually not prefer the type, but the
expr. Currently it seems like this can only happen with fwdrefs though,
which aren't really types anyways, and are probably rejected by the type
parsing in that case.

Doing sizeof() on fwdref's doesn't make sense, because they're not yet
resolved, but we don't show an error in that case, but just silently prefer
the var of the same name, and so the warning is probably useful even in
this case.