C backend: Prevent inline ASM keywords from being replaced by FB symbols
fixes #782

In inline ASM, we normally replace identifiers referring to some variable
or procedure by the mangled version, so a call to f becomes a call to F@0,
etc.

ASM keywords are explicitly excluded from such replacements, so it's
possible to have functions called "mov" etc. in the same source file that
also wants to use such keywords in inline ASM.

However, this was disabled for the C backend.

That's probably a left-over from when the C backend only supported the
gcc-style inline ASM syntax:
	asm
		"incl %0\n" : "+m" (a) : :
	end asm
where all inline ASM keywords are inside string literals, so they're not
parsed as identifiers anyways.

But the C backend also supports the classic FB inline ASM syntax nowadays:
	asm
		inc dword ptr [a]
	end asm
and here the keywords must be checked, as for the ASM backend.