internal: symbReuseOrAddConst(): Remove string debugging code
It was partially broken by c4a4dcfc, because now it will even be reached
in case the constant values are different, causing the assert()'s to fire
during the log-tests.

But these asserts() don't seem useful anymore anyways (no point checking
that the symbol has the same string data if we know it's the same symbol
anyways), so it's probably best to just remove them completely.