inc: winapi: Rename BOOLEAN => WINBOOLEAN, and maybe provide BOOLEAN
With the addition of a built-in boolean type to FB, it makes sense to
rename the BOOLEAN typedef from the WinAPI headers, to avoid collision.

This should mostly work, because FB's boolean and winapi's BOOLEAN will be
mostly compatible (at least they have the same size, and store 0/1 in
memory).

Renaming to WINBOOLEAN means this typedef can be used next to FB's boolean,
if wanted.

On the other hand, this patch also adds a block like this:
    #ifndef BOOLEAN
        type BOOLEAN as WINBOOLEAN
    #endif

which allows users to #undef FB's built-in boolean type before #including
the winapi headers, and then get BOOLEAN to refer to WINBOOLEAN, meaning
that old code using winapi's BOOLEAN will compile as before the addition
of the boolean type to FB.

This isn't 100% perfect, because existing code can be broken if users don't
add the #undef, but it's probably the best we can do, while adding the
built-in boolean to FB.

The alternative would be to have the WinAPI binding (and perhaps others)
silently override the built-in boolean type, making it impossible to use
FB's boolean once windows.bi was #included, or to trigger a compiler error
and force the user to explicitly decide everytime, but neither of these
seem like good ideas. By default, one would probably expect "boolean" to
refer to the language's built-in type, even if windows.bi was #included.