Adjust compiler to not rely on boolean support itself
This allows a fully working compiler to be bootstrapped using an older
version without boolean support. Previously the compiler relied on a
boolean-capable str() to fully work (but this only caused problems in the
test suite, not when rebuilding the compiler).

Besides that, it's also a chance to remove the use of cbool(). It's not
really needed anyways, since we don't use the boolean type anywhere in the
compiler code.