Change FileAttr() to return Integer and make fbFileAttrHandle work on 64bit
To support 64bit, FileAttr() needs to return an Integer instead of a Long,
because it wants to return pointers/handlers.

Internally, the old code dereferenced through the various internal structs
starting at FB_FILE.opaque. I've changed this to use the structs explicitly
instead of assuming int's everywhere. Just changing it to assume ssize_t's
everywhere would be wrong, because for example the LINUX_SERIAL_INFO stores
its OS handle (a file descriptor) as a 32bit int, even on 64bit.