Clean-up gfxlib during END instead of global dtor
(closes #750)

The gflxib needs to wait for its background thread to exit as part of its
clean-up. Previously this was done during the rtlib's global dtor.

However, waiting for a thread to exit during a global dtor is not safe on
Windows. There can be a dead-lock due to the loader lock that is acquired
when code such as DllMain() or global ctors/dtors is run. There is also a
race-condition there; if the thread finishes before the global dtor is
reached, then there's no dead lock.

Even though apparently the old code has worked fine for 10 years, it's
still not safe, and dead-lock problems were observed and reported on
Windows 10 (sometimes, not always).
http://www.freebasic.net/forum/viewtopic.php?f=6&t=23043

Hence the FB runtime needs to be fixed to clean-up the gfxlib during
fb_End(), i.e. during main(), instead of during a global dtor. For the
main supported use case this should not make a difference:
	screen 12
	print "hi"
	sleep

but there obviously is a difference for programs doing stuff like this:
* use FB graphics from global dtors - that won't work anymore
* using FB graphics inside DLLs - fb_End() isn't called here, the
  background will never terminate, and leave the program hanging.
  Such programs will have to be adjusted to do "screen 0" explicitly in
  the DLL before exiting.