inc: Turn many #defines into proper declarations in fbfrog-based bindings
The point here is to continue with the idea of avoiding #defines whereever
possible, because they are much more intrusive in the global namespace.

This is especially interesting for the Windows API binding:
	type FooA
	type FooW
	declare sub DoFooA
	declare sub DooFooW
	#ifdef UNICODE
		#define Foo FooW
		#define DoFoo DooFooW
	#else
		#define Foo FooA
		#define DoFoo DooFooA
	#endif
becomes:
	type FooA
	type FooW
	declare sub DoFooA
	declare sub DooFooW
	#ifdef UNICODE
		type Foo as FooW
		declare sub DoFoo alias "DooFooW"
	#else
		type Foo as FooA
		declare sub DoFoo alias "DooFooA"
	#endif
which frees up the identifiers Foo and DoFoo to be used in programs inside
namespaces even if the program #includes windows.bi. It's nice especially
for common identifiers like LoadImage.

This makes the Windows API binding more similar to fbc's old one again -
it also used alias declarations instead of defines, at least for the
A/W function aliases.