STABS emitter: Use "boolean" type tag for booleans, not "pchar"
The stabsTb() has an (apparently unused) entry that declares a type tag
with number 15 named "pchar". With boolean support came the new boolean
type tag with number 16. But the remapTB() entry for boolean referred to
15 instead of 16.

This fixes STABS data for boolean variables (i.e. they appear as booleans
now, not pchar anymore, in the debugger).