inc: Fix some broken translations in fbfrog bindings
Due to an fbfrog bug, some declarations were parsed and translated
incorrectly; it should be fixed now.

This problem showed up with the new jpeglib binding (added in the previous
commit) because sizeof(jpeg_decompress_struct) was completely wrong. The
old jpeglib binding was fine. (hence no changelog entry)

Apparently a struct in the gtk2 binding was also affected (even in the
older pre-1.02 GTK bindings), plus some probably irrelevant typedefs in
the winapi binding (that actually existed in the old winapi binding).