boolean: Ensure to promote integer operands even if the other is boolean
Previously, integer operand promotions were turned off in case of mixed
boolean/non-boolean operand. This left only the typeMax()-based
conversions, which isn't enough. For example, a BOP like "boolean and byte"
had typeof(byte), which seems wrong. The byte operand should be promoted
to integer, regardless of whether the other is a boolean or not.