Fix allocation of static member vars using the parent UDT as data type
related to bug #649

If a static member var using its parent UDT as data type is declared above
all or some of the UDT's fields, it is given the wrong size, because
sizeof(UDT) isn't fully known yet.

To fix this, such static member vars have to be updated (i.e. their size
must be recalculated) after the UDT was fully parsed, just like it was
already done with byval function results or byval parameters, if they
"forward-reference" the parent UDT.