C backend: Only use gcc inline asm constraints for variables
and not procedures/labels. Variables are memory, so it makes sense to use
the memory constraints for them, but procedures/labels are a different
story... they can't appear in =m constraints (gcc error: lvalue required).

There does not even seem to be any need to use any constraints for
procedures at all, things like "call f" can be written in gcc inline asm
too, as long as f really is the mangled/external name of the function.