int-only BOPs: Avoid truncating non-int operands
In a BOP like "integer xor double" we need to convert the double to an
integer because XOR can only operate on ints.

However, we might also have something like "longint xor double" - in this
case, previously the double was truncated to an integer, even though we
know it's supposed to be a longint operation. Now this is fixed, and the
non-int operand will be converted to an int type matching the other
operand.

"double xor double" still results in "integer xor integer" here though.

The issue with integer vs longint only matters on 32bit of course, but in
case of "uinteger xor double" we want the double to be converted to a
uinteger instead of just an integer, and this even matters for 64bit.