parser: Select Case As Const: Fix off-by-one when adding new Case values
The Select Case As Const parser stores Case values into an array with
FB_MAXSWTCASEEXPR slots. ctx.base represents the count of used slots.

The old code accessed ctx.base+1 which is not safe when we're adding the
last slot, because it refers to a next slot that doesn't exist. Generally
it always copied one slot too much. This isn't a big deal in practice
though because currently the Select Case As Const is limited by
FB_MAXSWTCASERANGE (4096), thus the FB_MAXSWTCASEEXPR (8192) limit can only
be reached during error recovery.