Fix PRINT comma padding wrapping into next line too early (#745)
I think the goal of this if check is to wrap into the next line if the
padding would move the cursor too far. But the old check was wrong; it did
not only wrap at position <maxwidth>, but already at
<maxwidth - FB_TAB_WIDTH> which surely isn't what we wanted.

This looks like a regression from a change to
FreeBASIC/src/rtlib/libfb_io_print.c in commit 659fe2763c5f.