Update cached screen width if graphics/console window is resized (#745)
We cache the screen width (in chars) in FB_HANDLE_SCREEN's width field (and
line_length is affected too). This is used by various things such as the
PRINT comma padding.

Previously the cached width was not updated when the screen size changed,
which can happen due to resizing the console window, or opening a new
graphics window.

Now I've added code for updating it when a new graphics window is opened
and when the Linux console window size changed, so things like
fb_PrintPadEx() get to see the proper values for the current screen.

I've also added the same to the Win32 console window size update function,
but it looks like that's only called during WIDTH, not when resizing the
window. The comment for fb_hUpdateConsoleWindow() explicitly says that's
how it must be, so I didn't change it any further. I don't know.