C backend: Fix dllimport'ed globals
dllimport (in FB's terms: Extern Import) means the symbol becomes a
pointer that the compiler automatically dereferences (like a byref
parameter).

fbc does this on the AST side already. Thus the C backend was broken, as it
used __attribute__((dllimport)). This meant that the implicit dereference
was done twice - once by fbc's AST and once by gcc.

The solution is to let the C backend emit dllimport variables just as the
ASM backend does it - as pointers, with the special _imp__ mangling prefix.