nsis-installer/makescript: Don't forget to clean up the last directory
Previously it only emitted rmdir commands for the previous file's directory
if the next file was in a different directory. This missed the case when
there is no next file.

However this didn't matter in practice because the last file always was
"readme.txt" which has no directories to clean up anyways.