Win32 rtlib: Fix double cursor position updating since 80b6026e
It seemed like a good idea to update FB_HANDLE_SCREEN's width and
line_length fields during the win32-specific fb_hUpdateConsoleWindow()
function, but at least with respect to line_length (the cursor position)
this is bogus, because to PRINT something we call fb_FilePutDataEx, which
1. calls fb_ConsolePrintBufferEx() which calls fb_hUpdateConsoleWindow()
   and thus updates the line_length
2. but also updates the line_length manually by doing line_length += foo.
That means the line_length was being updated twice, and ended up being
completely wrong (because step 2 assumed step 1 hasn't happened).

Thus it seems best to undo this change to the win32 part of the rtlib.