parser: Disallow fixed-length strings as BYREF vars/params/results
fixes part of #650

References or pointers to fixed-length strings are not possible in FB
because we have no way to encode the string length in the pointer dtype.
We can only store fixed-length string length on symbols, not on
expressions, which would be needed when dereferencing such a pointer.
Also, currently, taking the address of a fixed-length string produces an
"unbounded" Zstring Ptr.

Previously the string length was just ignored on BYREF symbols, so it seems
best to disallow it to avoid confusion, especially since it doesn't look
like it will be implemented properly soon (though of course that could
still be done).