parser: Fix/improve error recovery for procedure calls
This fixes a compiler crash during error recovery for BYREF AS ANY
parameters and (in theory) improves error recovery for procedure calls.

http://www.freebasic.net/forum/viewtopic.php?f=3&t=24140

The parser calls astNewCONSTz() to build some argument expression for
error recovery, but for ANY parameters, that returned NULL, causing
crashes later in astNewARG() which doesn't expect NULL.

Besides BYREF AS ANY parameters, it looks like astNewCONSTz() can only be
called with ANY type in case of SUBs (which internally use FB_DATATYPE_VOID
as type).

This patch changes the parser to build "proper" CALL expressions for error
recovery in the call parser, and it also changes astNewCONSTz() to never
return NULL anymore. If an ANY expression is requested, it just assumes
to be able to return an expression with any type.

This also means that cProcArgList() doesn't ever return NULL or a CONST
node anymore, but always a CALL node, which allows some further clean-ups.