Revert bad Get# wstring changes
In commit cab466533799 I changed Get#(wstring) to convert the loaded bytes
to wstring characters, but that's not how Get# should work:

It should just fill the buffer with the raw bytes. That's also what it
does for other data types like Integer variables - it will load
sizeof(integer) bytes, not just 1 byte.

This makes sense, since Get# is intended for use with Open For Binary.
The conversion to wstring chars is something that Input# with Open For
Input (i.e. text mode) would do.

That Get# change also broke part of fbc's UTF file reading, because it
uses Get#(wstring) (in hUTF16LEToUTF16LE() and hUTF32LEToUTF32LE()) and
relies on it to load raw bytes as-is without any conversion.