Eof: Don't check ftell() in text-mode
ftell() and ftello64() cannot safely be used on text-mode streams in the
way Eof() was using it: It assumed the returned value was a valid file
offset corresponding to the current position, but ftell() isn't guaranteed
to actually return that. It's only guaranteed to return something that can
be used with fseek().

I ran into this issue on Windows when reading a file with LF line endings
(instead of CRLF) in text mode. The ftell() value was different from the
real file offset (probably due to the internal LF => CRLF conversions?),
causing Eof() to return TRUE too early. I've added the test case - it
failed for me on Windows.

It seems that whether ftell() gives a useful result also depends on file
size and the buffer size set with setvbuf(... _IOFBF ...) (FB uses that
too). The point is that ftell() isn't guaranteed to give valid file offsets
in text-mode.

Instead, this patch changes Eof() to peek ahead via getc()/ungetc(), and
check for EOF this way.