rtlib/gfxlib2: Fix RORW() macro for non-x86
It didn't truncate the value to 16bit, causing it to fail when working
with 32bit variables, which is how the gfxlib2 box/line drawing code uses
it. Sooner or later it would start shifting the upper 16bits into the lower
16bit, and eventually end up with all 1's, meaning the <if (style & bit)>
type of checks used by fb_GfxDrawLine() etc. would always be true. Thus it
eventually started drawing all pixels instead of the wanted pattern.

http://www.freebasic.net/forum/viewtopic.php?p=214164#p214164