libfbdoc: Add support for CSRFToken for Wikka login
When logging in to the Wikka wiki, we have to send the CSRFToken value as
part of the form POST. This CSRFToken is embedded in the HTML page
containing the login form (UserSettings), from where we can extract it.
Without this, the wiki refuses login or uploads with an "authentication
error".