wiki: Fix some cases where the auto-conversion misdetected links
Looks like two wiki pages have instances of [[ that however do not
represent links, confusing the updatelinks tool.

I noticed this because on these pages there are secondary links that did
not get converted, e.g.:
	[[not a link...   [[but this is a link]]
became:
	[[not|a link...   [[but this is a link]]
while it should be:
	[[not a link...   [[but|this is a link]]
and the unconverted links caused fbdoc to show errors when generating
the offline docs.

The updatelinks tool itself isn't really needed anymore, so it can aswell
be removed again.