Fix CONSTness check for virtual dtor calls (#732)
Virtual destructors couldn't be called on const objects anymore, due to
f4c8b942 which changed the non-const-method-called-on-const-object check
to explicitly exclude ctors/dtors. However, that fails for virtuals,
because the vtable access procptr is checked, not the original proc,
and the procptr doesn't have the ctor/dtor attrib.

To fix this, this adds a new attrib to mark ctors/dtors and procptrs to
them as special, so that astNewARG() can disable its THIS CONSTness check.

Currently it's not possible to just propagate the ctor/dtor (and method/
property/operator) flags to procptr types (similar to d54173d7), because
hSetupProc() and possibly other code assumes any PROC symbol with those
attribs is a real method, not just a procptr subtype.

This fix would not be necessary if the type checks would be done on the
original called proc symbol, instead of on the vtable access procptr as
it is done currently. But for that, either astNewCALL() must be provided
with the additional information, or virtualization must be delayed until
after astNewCALL() (so the ARG type checking never gets to see the vtable
access procptr).