Fix compiler crash after REDIM error
addsuffix could be set to TRUE even when there was no suffix, which could
cause array symbols created after REDIM error message to be marked as
SUFFIXED even though they could have a type that can't be a suffix.

REDIM is the only case where dtype can be omitted, so it's the only case
where addsuffix wasn't always overridden by dtype parsing. And if it's a
typeless REDIM then it works on an existing symbol instead of creating a
new one, so addsuffix would be ignored anyways - except if there's an error
and a new symbol is created during error recovery.