Unix rtlib: Fix exec() to support spaces in executable path
fb_hGetShortPath() escaped spaces as "\ " in paths containing spaces,
which is bad here because the path is passed to execvp() which is not
a shell but just expects the raw file name.

On top of that, the fb_hConvertPath() done afterwards converts \ to /,
so spaces were turned into "/ ", resulting in a completely different path
than what was given to exec().

This fixes the former, so the second won't kick in in this case either
anymore, but the \ => / conversion in general is not removed here. That's
another discussion...