Fix code generation for SELECT CONST with CASE -1 on 64bit
fbc used a For loop like "For value As ulongint = minval To maxval",
but this fails (infinite loop) in case maxval = &hFFFFFFFFFFFFFFFFull,
which happens when for example Case -1 is used in the Select Case As Const.

The bad code existed in both ASM and C backends, but since it only affects
64bit (because on 32bit, the Case values are truncated to 32bit UInteger,
which fits into ULongInt just fine), the ASM backend doesn't trigger the
issues.

http://www.freebasic.net/forum/viewtopic.php?p=225109#p225109