Ensure to truncate 32bit jump table index to 32bit
astBuildJMPTB() calculated -minval * env.pointersize with
minval=&hFFFFFFFFll, resulting in &hFFFFFFFC00000004. This must be
truncated to 32bit to get the wanted result of 4. The value is later
assigned to some Integer in emit_x86.bas or somewhere else, so it happened
to work in 32bit fbc, but not in 64bit fbc.

The solution is to ensure to truncate this precomputed index offset to
32bit in case of 32bit Select Case As Const.

Ideally the ASM backend would behave the same for 32bit and 64bit (i.e. use
LongInts instead of Integers), but that's a different issue.

Test case for 64bit fbc cross-compiling to 32bit:
  select case as const 0u
  case -1
  end select

  jmp dword ptr [.LT_0008+eax*4-17179869180]

  1.asm:31: Error: value of -17179869116 too large for field of 4 bytes at 60