Fix surrogate pair truncation in UTF32 -> Windows WString conversion
Make the UTF32 -> Windows WString case consistent with other UTF32 -> UTF16
cases in FB rtlib: When the destination buffer is too small to hold a
surrogate pair, only the low (second) surrogate is written, and the high
(first) surrogate is omitted.

This is pretty strange (producing invalid UTF16 instead of failing),
so I don't want to document this in detail (it should probably be changed),
but at least FB's UTF16-generation cases are now consistent with regards
to the handling of truncating surrogate pairs.

Previously this code just wrote the low 16bit of the code point, which is
invalid aswell.