Fix compiler crash on type<>() initializer for non-UDT global var
1. astTypeIniIsConst/hExprIsConst wanted to do checks against the global
var symbol when checking a global var init, however the TYPEINI sym can be
NULL in case of something like type<integer>(0), so it crashed. To fix it,
I moved the checks out of astTypeIniIsConst. astLoadStaticInitializer/
hFlushExprStatic had a similar problem, but since the basesym is passed
there next to the TYPEINI, it's easier to fix.
(Normally a TYPEINI like type<integer>(0) would have a temp var sym, but
global var inits aren't processed by astTypeIniUpdate.)

2. Refactor code in parser-decl-var.bas to use same global var init
validation checks for both Dim and Var.