Fix #851: #line directive doesn't change source filename in debug info
Added a new hash table, env.filenamehash, for all input filenames;
env.inconcehash and env.incfilehash are now subsets of this hash.

The name of the main source file is also stored here, so env.inf.incfile
is now never NULL (allowing a little cleanup in ir-hlc.bas).
Really, env.inf.incfile is misnamed, as it is used just as an interned
copy of env.inf.name so that it can be stored in the debug info for
a procedure for use later. env.inf.name could be removed and its use
replaced with env.inf.incfile.