C backend: fix #858 - internal structure size mismatch due to padding differences
A structure which decreases its alignment (using FIELD = n) is annotated with both the "packed" and the "alignment(n)" attributes. That leads to a problem if such a structure is nested inside another structure that uses a greater alignment than the child structure. In that case the field representing the nested structure is aligned with the same properties as the nested structure.

So the alignment has to be increased explicitly if the nested structure is packed with a smaller alignment than the parent structure. That is accomplished by adding the "alignment(n)" attribute to those fields.