Fix emitting of CONST ref initializers (as opposed to OFFSETs)
The CONST code path in hFlushExprStatic() used the symbol type as-is,
instead of converting to pointer type for references. Thus the type
looked different and it attempted the astNewCONV() which could easily
fail (because the init expression is a pointer, while the symbol could
be anything, e.g. struct/string).