Fix mangling involving toplevel CONST and abbreviations
This attempts to fix a mangling problem with `BYVAL x AS CONST TYPE`
parameters. A toplevel CONST like that should be ignored when mangling,
just like it's ignored in a C++ function signature. fbc already
correctly skipped the `K`, but it should also not add an abbreviation
for this case, to be compatible with gcc.

Checking gcc (on x86_64 linux):
```
void f(      int, int &, int &) { } // _Z1fiRiS_
void g(const int, int &, int &) { } // _Z1giRiS_
```
No `K` added, and the abbreviation used in the 3rd parameter is `S_`,
which is the first abbreviation, meaning that none was added for the
first parameter. Also, note that C++ considers both signatures equal.

https://www.freebasic.net/forum/viewtopic.php?t=31697