make log-tests: Treat fbc crash as error for COMPILE_ONLY_FAIL tests
COMPILE_ONLY_FAIL tests test that fbc prints an error message and fails.
However, it was not able to differentiate an intended fbc error from
an unintended fbc crash.

This patch adds check of the fbc exit code, which catches some of the
unexpected errors, most notably SIGSEGV crashes. It also catches FB
runtime errors such as "null pointer access", if fbc was built with
-exx. However, it cannot detect "illegal function call", because that
uses value 1, which is also the exit code for expected errors.