rtlib: unix: Fix tty state capture logic and FILE* leak
fb_hInitConsole() and fb_hExitConsole() did not handle unbalanced
calls correctly. If fb_hInitConsole() was called twice in a row without
an fb_hExitConsole() in between, it simply overwrote its internal state,
leading to the FILE*/fd leak (from the fopen("/dev/tty")),
and also leading to it losing track of the original in/out tty state
captured at start-up. Instead, the secondary invocations of tcgetattr()
captured the tty states configured by the FB runtime itself during the
previous fb_hInitConsole() call.
Thus, fb_hExitConsole() could not restore the original state on exit.

The most noticable symptom of this for me was that echoing remained
disabled in the terminal even after the FB program exited, with programs
such as this:
	sub threadmain(byval arg as any ptr)
		shell("echo thread; sleep 1")
	end sub
	var thread = threadcreate(@threadmain)
	shell("echo main; sleep 0.5")
	threadwait(thread)
This case is also covered by the tests in tests/unix-tty/, which now
succeeds.

Why can the unbalanced calls happen? The affected functions are
fb_DylibLoad(), fb_DylibFree(), fb_ExecEx(), fb_hShell().
They use fb_hExitConsole() and fb_hInitConsole() to temporarily
undo the FB runtime's tty configuration, so that an FB runtime (if any)
in the sub-process or loaded shared library can get to see and capture
it, for proper restoring if it exits after the parent process (and
perhaps for more reasons?). See commits 9599d6d060, 1be51c1f86.

While these functions ensure basic thread-safety by wrapping each call
to fb_hExitConsole() or fb_hInitConsole() in an FB_LOCK()/FB_UNLOCK()
since commit 6b4dfd8623, they only lock the lock for the duration of
the single call, not both calls together, because then the lock would
be held for the entire duration of a shell()/exec() call, which would
not be desirable.
Because of the incomplete locking, it's easily possible for calls to
fb_hInitConsole() and fb_hExitConsole() to be unbalanced in scenarios
with multiple threads calling the affected functions in parallel.