rtlib: unix: Drop support for old libgpm versions (libgpm.so.1)
It seems useless to keep supporting the old versions, plus there are
ABI incompatibility problems if the libgpm version used when compiling
the rtlib is not the same as the one used at runtime. The easiest
solution seems to be to just drop support for the old versions.

Otherwise we'd have to try to detect the version at runtime,
and use the proper offsets based on that...