Add test program for test of Unix rtlib TTY state handling
This includes a tester program and various testee programs, for testing
the behaviour of fb_hInitConsole()/fb_hExitConsole() of the Unix rtlib.
See the readme.txt etc. for details.

Currently this fails for the parallel shell() test, due to a bug in the
rtlib, which will be fixed in the next commit. That's the kind of issue
I want to find with this test program.

I'm not sure yet how to best integrate this test program. These tests
apply to Unix systems only. I don't think they can be integrated into
the existing cunit/fbcunit tests, log-tests or warning-tests.
Thus, I've made it completely stand-alone for now.