fbc: multithread programs and fbgfxmt linking
This fixes a few issues when using fbgfx and threaded (-mt) programs.

The main problem was that if "fbgfx.bi" is included and threading is expected
either with -mt command line switch or use of any threading functions, then
we get either -lfbgfx or both -lfbgfx and -lfbgfxmt passed to the linker where
we would expect only the thread safe version of fb gfx library.

   program.bas
      #include "fbgfx.bi"
      sub proc( byval arg as any ptr )
      end sub
      threadcreate( @proc, 0 )
      screen 12

   $ fbc -v [-mt] program.bas
     linking: ld ... -lfbgfx -lfbmt -lfbgfxmt ...

   Or whithout the graphics statements:
   $ fbc -v [-mt] program.bas
     linking: ld ... -lfbgfx -lfbmt ... ...

Changes to help ensure that only one correct fb gfx library is passed to the linker:

- fbc will restart the parser when #cmdline "-mt" is used or usage of any threaded
  functions automatically enable -mt.  This allows the compiler to parse the current
  module from the beginning and process any __FB_MT__ checks.
- Add special handling for #inclib "fbgfx?" to fix incorrect linking to non-threaded
  gfx library.  #include "fbgfx.bi" was causing the libfbgfx library to always be
  linked even when thread safe version was expected (causing deadlocks in user
  programs)